backend/pyproject.toml
[project]
name = "drawing-tutor-backend"
version = "0.1.0"
description = "Backend API for Drawing Tutor application"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "motor>=3.3.2",
    "pymongo>=4.6.0",
    "pillow>=10.2.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "aiofiles>=23.2.1",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.26.0",
]

[tool.uv]
dev-dependencies = []